requirements.txt
argcomplete==1.12.1
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Django==2.1.2
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
jmespath==0.10.0
kappa==0.6.0
mysqlclient==2.0.1
pip-tools==5.3.1
placebo==0.9.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.2
troposphere==2.6.2
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.3.0
argcomplete==1.12.1
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Django==2.1.2
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
jmespath==0.10.0
kappa==0.6.0
mysqlclient==2.0.1
pip-tools==5.3.1
placebo==0.9.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.2
troposphere==2.6.2
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.ㅎ6
zappa==0.51.0
zipp==3.3.0
argcomplete==1.12.1
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Django==3.1.2
django-extensions==3.0.9
django-storages==1.9.1
djangorestframework==3.12.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
jmespath==0.10.0
kappa==0.6.0
mysqlclient==2.0.1
pip-tools==5.3.1
placebo==0.9.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.2
troposphere==2.6.2
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.3.0
argcomplete==1.12.1
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-extensions==3.0.9
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.12.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
pip-tools==5.3.1
placebo==0.9.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.2
troposphere==2.6.2
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.3.0
